move rendered_activity function component to ActivityHelpers module
